# swagger.yaml
openapi: 3.0.0
info:
  title: Pacific Telecom
  description: Apis para Pacific Telecom
  version: 0.1
servers:
  # - url: http://localhost:3000
  - url: http://138.197.105.129:3000
    description: Pacific Telecom Resource Server

# 1) Define the security scheme type (HTTP bearer)
components:
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes
# 2) Apply the security globally to all operations
security:
  - bearerAuth: []         # use the same name as above
paths:
...
  /plans/{id}:
    get:
      summary: Obtener un plan por id
      tags:
        - Planes
      parameters:
        - name: id
          in: path
          required: true
          description: id del plan
          schema:
            type : integer
            format: int64
            minimum: 1
      responses: 
        '200':
          description: OK

    /plans/{id}:
    delete:
      summary: Borra un plan por id
      tags:
        - Planes
      parameters:
        - name: id
          in: path
          required: true
          description: id del plan
          schema:
            type : integer
            format: int64
            minimum: 1
      responses: 
        '200':
          description: OK

      /plans/{id}:
    patch: 
      summary: Actualizar un plan por id
      tags:
        - Planes
      parameters:
        - name: id
          in: path
          required: true
          description: id del plan
          schema:
            type : integer
            format: int64
            minimum: 1
      requestBody:   
        content:
          application/json:
            schema:
              type: object
            example: '{"name": "Conexion fibra optica","description": "Conexion de alta rapidez","monthly_price": 100000000,"long_term_price":10000000, "long_term_type":1,"service_id": 1,"weight": 1,"mbps_up":1, "mpbs_down": 3, "reuse_factor":5, "max_limit_up":2,"limit_at_up":2,"max_limit_down":2, "limit_at_down":2, "burst_limit_upload":2, "burst_limit_download":2, "burst_threshold_upload":2, "burst_threshold_download":2,"burst_time":2, "priority":2, "parent":1, "max_device_public_internet":2, "max_limit_value":2,"bw_required":2}'
      responses: 
        '200':
          description: OK
  /plans/register:
    post: 
      summary: Registrar plan
      tags:
        - Planes
      requestBody:   
        content:
          application/json:
            schema:
              type: object
            example: '{"name": "Conexion fibra optica","description": "Conexion de alta rapidez","monthly_price": 100000000,"long_term_price": 10000000, "long_term_type": 1,"service_id": 1,"mbps_up": 1,"mpbs_down": 1,"max_limit_parent_up": 1,"max_limit_parent_down": 1,"reuse_factor": 1,"max_limit_up": 1,"limit_at_up": 1,"max_limit_down": 1, "limit_at_down": 1,"burst_limit_upload": 1,"burst_limit_download": 1,"burst_threshold_upload": 1,"burst_threshold_download": 1,"burst_time": 1,"burst": false,"burst_x100": 1,"priority": 1,"parent": "DOWNLOAD","max_device_public_internet": 1,"max_limit_value": 1,"bw_required": 1,"weight": 1,"bucket": false,"active_service_quantity": 1}'
      responses: 
        '201':
          description: OK

############################################### [URBANISMOS] ######################################
...